🎨 Clearly defined the OutputConnection <continued>
Previously there existed the "BasicConnection", which was an incredibly ambigious name. It has been replaced by the "OutputConnection" which represents a connection to an output from another node.

The other types of connections that exist are the ConnectionTypeConverter and the static value connection which has yet to be implemented fully.